Catch more misaligned pointers via careful padding
The current pointer alignment adds a random amount of padding
between allocations (possibly). That has at best a 50% chance of
detecting a failure to check alignment between types whose alignment
only differs by a factor of 2, and because it is random, crates with a
small number of tests can slip by entirely.

This changes to a deterministic algorithm which always detects an
unchecked access of a pointer to T as a pointer to another type with
twice T's alignment.