DO NOT MERGE - scratchwork
This shows how the base address, if it can be acquired, can be used to
call external C functions. There are numerous TODOs, including the big
one of coming up with a way to access the bytes address without the huge
hole punch I put in rustc.

This also requires that the bytes addresses in rustc become properly
aligned.

Double dereference passes, as do most tests. Of the few that remain,
some of them make questionable assumptions (which we might need some way
to retain or put behind an option for sanitization) like that it is
fundamentally possible for two allocations to end up next to each other
- this won't happen when we're using a real heap. The simd tests failing
  are concerning, but can be looked into.