Implement support for some basic wasm32-wasi syscalls
This allows a basic "hello world" to run, as well as
`cargo miri test --target wasm32-wasi`

Wasi is a somewhat unusual platform. Rust doesn't consider
it to be unix-like (std::os::unix is not available, and the target spec
doesnt have "unix" as a target family). However, it has a partial unix
libc implementation based on wasi-specific syscalls
(https://github.com/WebAssembly/wasi-libc), and libstd relies on this
in order to call libc functions in wasi-specific code.

As a reuslt, I've decided to mark wasi as a unix platform in Miri.
This lets us re-use all of the unix libc function implementations,
which are needed for things like environment variables to work properly.

wasm32-wasi currently lacks support for threads, though a proposal
for thread support is being actively developed
(https://github.com/WebAssembly/wasi-threads).
As a result, I've had to disable a very large number of tests which
rely on on thread support. While this requires editing a large number
of files, we still run the majority of tests under wasm32-wasi.
Assuming that wasi eventually gains thread support, we can re-enable
these tests on wasi at some point in the future.

Some other wasm32-wasi limitations:
* The wasm32 target only supports panic=abort. While Miri could
  technically ignore this limitation and attempt to unwind anyway,
  I suspect that this could have weird consequences. As a result,
  I've ignored any tests that depend on `catch_unwind`
* Unfortunately, the current version of the `wasi` syscall crate
  passes all pointers arguments by first performing a ptr-to-integer
  cast. As a result, virtually any wasm32-wasi program will trigger
  Miri provenance warnings by default. When running the test suite
  against wasm32-wasi, I've set MIRIFLAGS="-Zmiri-permissive-provenance"
  which allows almost all tests to pass. There are a few tests which
  specifically rely on the *default* provenance setting
  (they expect provenance warnings in the test stderr). Since we
  currently lack a way to reset the provenance setting back to the
  default, I've disabled these tests on wasm.