README: Mention test-tag
The usage of `#[cfg_attr(miri, ignore)]` works well if the majority of
tests is Miri compatible, but can become cumbersome if that's not the
case. The `test-tag` crate provides the means for having an opt-in
scheme instead, which may be preferred in larger code bases.

Given that the README already mentions `nextest` as a third party crate,
embed a short note about the possibility of using `test-tag` as well.