[WIP] Use the dummy codegen backend
This allows getting rid of all the argument mangling hacks as well as
the exported_non_generic_symbols override hack by doing cargo build
instead of cargo check. This is also a prerequisite for using native jit
mode support in miri that I hope to add to rustc in the future to be
used with cg_clif too.

Currently this still has two issues: The dummy backend can't build
dylibs, but cargo tries to build libstd as dylib. And required target
features are not set by the dummy backend. Both of these require rustc
side changes.