Restore support for the wasm32-wasip2 target
This adjusts find_mir_or_eval_fn to respect #[wasm_import_module] on
wasm. To avoid having to thread through the import module everywhere, I
opted to merge it with the item name with a $$$ as separator. This is
unlikely to occur in a real import module name or item name.

This reverts commit c38128b0d43ae39f93bc8039b7950ab2000ec925 and makes
a few minor adjustments.