Basic support for FFI callbacks
This commit adds basic support for FFI callbacks by registering a shim
function via libffi. This shim function currently only notes if it's
actually called and registers an error for these cases. The main
motivation for this is to prevent miri segfaulting as described in
[4639](https://github.com/rust-lang/miri/issues/4639).

In the future miri could try to continue execution in the registered
callback, although as far as I understand Ralf that is no easy problem.